+++ webpack.config.js
@@ -1,5 +1,6 @@
 const NodePolyfillPlugin = require("node-polyfill-webpack-plugin")
 const path = require('path');
+const webpack = require('webpack');
 
 module.exports = {
     mode: "production",
@@ -13,6 +14,9 @@ module.exports = {
     },
     target: ['web', 'es5'],
     plugins: [
-        new NodePolyfillPlugin()
+        new NodePolyfillPlugin(),
+        new webpack.DefinePlugin({
+            CLIENT_ID: JSON.stringify("https://knoodle.knows.idlab.ugent.be/id")
+        })
     ]
 };